requirements.txt
Appium-Python-Client==0.52;python_version < '3.0'
Appium-Python-Client==1.0.2;python_version >= '3.0'
Appium-Python-Client~=1.0.2
selenium~=3.141.0
pytest
allure-pytest
requests~=2.28.1
Python-BrowserStack~=1.2